Fixed: Improve moving file to location where another one exists

+++ src/NzbDrone.Common/Disk/DiskProviderBase.cs
@@ -265,6 +265,11 @@ namespace NzbDrone.Common.Disk
 
         protected virtual void MoveFileInternal(string source, string destination)
         {
+            if (File.Exists(destination))
+            {
+                throw new FileAlreadyExistsException("File already exists", destination);
+            }
+
             File.Move(source, destination);
         }
